# frontend/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sales-Assistent - Mail-Modul",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/drafts": {
      "get": {
        "tags": [
          "drafts"
        ],
        "summary": "List Drafts",
        "operationId": "list_drafts_api_drafts_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DraftStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DraftOut"
                  },
                  "title": "Response List Drafts Api Drafts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}": {
      "get": {
        "tags": [
          "drafts"
        ],
        "summary": "Get Draft",
        "operationId": "get_draft_api_drafts__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "drafts"
        ],
        "summary": "Update Draft",
        "operationId": "update_draft_api_drafts__draft_id__put",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/approve": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Approve Draft",
        "description": "Sends an approved draft, exactly once.\n\nTwo phases, because sending is irreversible:\n\n1. Claim. The draft row is locked, checked, moved to FREIGEGEBEN and\n   committed. The lock is released immediately; the claim is durable.\n   A second, concurrent approval now blocks on the lock, then sees a\n   status that is no longer ENTWURF and is rejected - where previously\n   both requests passed the check and the customer got the mail twice.\n2. Send, then record. Only after Gmail accepts does the draft move to\n   VERSENDET and the mail to ERLEDIGT.\n\nIf the process dies between the two phases the draft stays in\nFREIGEGEBEN, visible and recoverable by a human. That is the\ndeliberate direction of the trade: a stuck draft can be fixed, a\nduplicate mail to a customer cannot be un-sent.",
        "operationId": "approve_draft_api_drafts__draft_id__approve_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/drafts/{draft_id}/reject": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Reject Draft",
        "operationId": "reject_draft_api_drafts__draft_id__reject_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRejectIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "summary",
          "status",
          "created_at"
        ],
        "title": "CaseOut"
      },
      "CaseStatus": {
        "type": "string",
        "enum": [
          "offen",
          "geschlossen"
        ],
        "title": "CaseStatus"
      },
      "ContactOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email_address": {
            "type": "string",
            "title": "Email Address"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email_address",
          "name",
          "company"
        ],
        "title": "ContactOut"
      },
      "DraftOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email_message_id": {
            "type": "string",
            "format": "uuid",
            "title": "Email Message Id"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "status": {
            "$ref": "#/components/schemas/DraftStatus"
          },
          "rag_context_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rag Context Summary"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "sent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "email_message": {
            "$ref": "#/components/schemas/EmailSummaryOut"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email_message_id",
          "subject",
          "body",
          "status",
          "rag_context_summary",
          "created_at",
          "updated_at",
          "sent_at",
          "email_message"
        ],
        "title": "DraftOut"
      },
      "DraftRejectIn": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "DraftRejectIn"
      },
      "DraftStatus": {
        "type": "string",
        "enum": [
          "entwurf",
          "freigegeben",
          "abgelehnt",
          "versendet"
        ],
        "title": "DraftStatus"
      },
      "DraftUpdateIn": {
        "properties": {
          "subject": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 998
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "DraftUpdateIn"
      },
...
      "EmailStatus": {
        "type": "string",
        "enum": [
          "neu",
          "abgelegt",
          "wartet_auf_freigabe",
          "erledigt",
          "ausgeblendet"
        ],
        "title": "EmailStatus"
      },
      "EmailSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "sender_address": {
            "type": "string",
            "title": "Sender Address"
          },
          "sender_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Name"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "wichtigkeits_kategorie": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WichtigkeitsKategorie"
              },
              {
                "type": "null"
              }
            ]
          },
          "typ": {
            "$ref": "#/components/schemas/TypKategorie"
          },
          "status": {
            "$ref": "#/components/schemas/EmailStatus"
          },
          "contact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContactOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "case": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject",
          "sender_address",
          "sender_name",
          "snippet",
          "received_at",
          "wichtigkeits_kategorie",
          "typ",
          "status",
          "contact",
          "case"
        ],
        "title": "EmailSummaryOut",
        "description": "Everything a list view needs - deliberately without `raw_content`,\nwhich is the largest field by far and never rendered in a list."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TypKategorie": {
        "type": "string",
        "enum": [
          "bestellung",
          "anfrage",
          "keiner"
        ],
        "title": "TypKategorie"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WichtigkeitsKategorie": {
        "type": "string",
        "enum": [
          "antwort_erforderlich",
          "information",
          "newsletter",
          "spam_verdacht"
        ],
        "title": "WichtigkeitsKategorie"
      }
    }
  }
}